LPS-107004 Break into Process and ProcessMetric
