+++ modules/dxp/apps/portal-workflow/portal-workflow-metrics-rest-impl/src/main/java/com/liferay/portal/workflow/metrics/rest/internal/resource/v1_0/ProcessResourceImpl.java
@@ -16,22 +16,6 @@ package com.liferay.portal.workflow.metrics.rest.internal.resource.v1_0;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.search.Field;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.odata.entity.EntityModel;
-import com.liferay.portal.search.aggregation.AggregationResult;
-import com.liferay.portal.search.aggregation.Aggregations;
-import com.liferay.portal.search.aggregation.bucket.Bucket;
-import com.liferay.portal.search.aggregation.bucket.FilterAggregation;
-import com.liferay.portal.search.aggregation.bucket.FilterAggregationResult;
-import com.liferay.portal.search.aggregation.bucket.TermsAggregation;
-import com.liferay.portal.search.aggregation.bucket.TermsAggregationResult;
-import com.liferay.portal.search.aggregation.metrics.ScriptedMetricAggregationResult;
-import com.liferay.portal.search.aggregation.metrics.ValueCountAggregationResult;
-import com.liferay.portal.search.aggregation.pipeline.BucketSelectorPipelineAggregation;
-import com.liferay.portal.search.document.Document;
 import com.liferay.portal.search.engine.adapter.search.SearchRequestExecutor;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchRequest;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchResponse;
@@ -39,32 +23,18 @@ import com.liferay.portal.search.hits.SearchHit;
 import com.liferay.portal.search.hits.SearchHits;
 import com.liferay.portal.search.query.BooleanQuery;
 import com.liferay.portal.search.query.Queries;
-import com.liferay.portal.search.query.StringQuery;
-import com.liferay.portal.search.query.TermsQuery;
-import com.liferay.portal.search.script.Scripts;
-import com.liferay.portal.search.sort.FieldSort;
-import com.liferay.portal.search.sort.SortOrder;
-import com.liferay.portal.search.sort.Sorts;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-import com.liferay.portal.vulcan.resource.EntityModelResource;
+import com.liferay.portal.vulcan.util.LocalizedMapUtil;
 import com.liferay.portal.workflow.metrics.rest.dto.v1_0.Process;
-import com.liferay.portal.workflow.metrics.rest.internal.odata.entity.v1_0.ProcessEntityModel;
-import com.liferay.portal.workflow.metrics.rest.internal.resource.helper.ResourceHelper;
+import com.liferay.portal.workflow.metrics.rest.exception.v1_0.NoSuchProcessException;
+import com.liferay.portal.workflow.metrics.rest.internal.dto.v1_0.util.ProcessUtil;
 import com.liferay.portal.workflow.metrics.rest.resource.v1_0.ProcessResource;
+import com.liferay.portal.workflow.metrics.search.index.ProcessWorkflowMetricsIndexer;
 import com.liferay.portal.workflow.metrics.search.index.name.WorkflowMetricsIndexNameBuilder;
 
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Date;
-import java.util.LinkedHashMap;
-import java.util.LinkedList;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Stream;
-
-import javax.ws.rs.core.MultivaluedMap;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -75,26 +45,29 @@ import org.osgi.service.component.annotations.ServiceScope;
  */
 @Component(
 	properties = "OSGI-INF/liferay/rest/v1_0/process.properties",
-	scope = ServiceScope.PROTOTYPE,
-	service = {EntityModelResource.class, ProcessResource.class}
+	scope = ServiceScope.PROTOTYPE, service = ProcessResource.class
 )
-public class ProcessResourceImpl
-	extends BaseProcessResourceImpl implements EntityModelResource {
+public class ProcessResourceImpl extends BaseProcessResourceImpl {
 
 	@Override
-	public EntityModel getEntityModel(MultivaluedMap multivaluedMap)
-		throws Exception {
+	public void deleteProcess(Long processId) throws Exception {
+		Process process = getProcess(processId);
 
-		return _entityModel;
+		_processWorkflowMetricsIndexer.deleteProcess(
+			contextCompany.getCompanyId(), process.getId());
 	}
 
 	@Override
-	public Process getProcess(
-			Long processId, Boolean completed, Date dateEnd, Date dateStart)
-		throws Exception {
+	public Process getProcess(Long processId) throws Exception {
+		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
+
+		searchSearchRequest.setIndexNames(
+			_processWorkflowMetricsIndexNameBuilder.getIndexName(
+				contextCompany.getCompanyId()));
+		searchSearchRequest.setQuery(_createBooleanQuery(processId));
 
 		return Stream.of(
-			_getProcessesSearchSearchResponse(null, null, processId, null)
+			_searchRequestExecutor.executeSearchRequest(searchSearchRequest)
 		).map(
 			SearchSearchResponse::getSearchHits
 		).map(
@@ -105,53 +78,27 @@ public class ProcessResourceImpl
 			SearchHit::getDocument
 		).findFirst(
 		).map(
-			document -> {
-				Process process = _createProcess(document);
-
-				Bucket bucket = _getProcessBucket(
-					GetterUtil.getBoolean(completed), dateEnd, dateStart,
-					processId);
-
-				_populateProcessWithSLAMetrics(bucket, process);
-				_setInstanceCount(bucket, process);
-
-				_setUntrackedInstanceCount(process);
-
-				return process;
-			}
-		).orElseGet(
-			Process::new
+			document -> ProcessUtil.toProcess(
+				document, contextAcceptLanguage.getPreferredLocale())
+		).orElseThrow(
+			() -> new NoSuchProcessException(
+				"No Process exists with the id " + processId)
 		);
 	}
 
 	@Override
-	public Page<Process> getProcessesPage(
-			String title, Pagination pagination, Sort[] sorts)
-		throws Exception {
-
-		FieldSort fieldSort = _toFieldSort(sorts);
-
-		SearchSearchResponse searchSearchResponse =
-			_getProcessesSearchSearchResponse(
-				fieldSort, pagination, null, title);
-
-		long count = searchSearchResponse.getCount();
-
-		if (count > 0) {
-			return Page.of(
-				_getProcesses(
-					fieldSort, pagination,
-					searchSearchResponse.getSearchHits()),
-				pagination, count);
-		}
-
-		return Page.of(Collections.emptyList());
-	}
-
-	@Override
 	public String getProcessTitle(Long processId) throws Exception {
+		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
+
+		searchSearchRequest.setIndexNames(
+			_processWorkflowMetricsIndexNameBuilder.getIndexName(
+				contextCompany.getCompanyId()));
+		searchSearchRequest.setQuery(_createBooleanQuery(processId));
+		searchSearchRequest.setSelectedFieldNames(
+			"processId", _getTitleFieldName());
+
 		return Stream.of(
-			_getProcessesSearchSearchResponse(null, null, processId, null)
+			_searchRequestExecutor.executeSearchRequest(searchSearchRequest)
 		).map(
 			SearchSearchResponse::getSearchHits
 		).map(
@@ -168,438 +115,42 @@ public class ProcessResourceImpl
 		);
 	}
 
-	private BooleanQuery _createBooleanQuery(boolean completed) {
+	@Override
+	public Process postProcess(Process process) throws Exception {
+		return ProcessUtil.toProcess(
+			_processWorkflowMetricsIndexer.addProcess(
+				process.getActive(), contextCompany.getCompanyId(),
+				process.getDateModified(), process.getDescription(),
+				process.getDateModified(), process.getName(), process.getId(),
+				process.getTitle(),
+				LocalizedMapUtil.getLocalizedMap(process.getTitle_i18n()),
+				process.getVersion()),
+			contextAcceptLanguage.getPreferredLocale());
+	}
+
+	@Override
+	public void putProcess(Long processId, Process process) throws Exception {
+		Process getProcess = getProcess(processId);
+
+		Map<Locale, String> titleMap = LocalizedMapUtil.getLocalizedMap(
+			process.getTitle_i18n());
+
+		_processWorkflowMetricsIndexer.updateProcess(
+			process.getActive(), contextCompany.getCompanyId(),
+			process.getDescription(), process.getDateModified(),
+			getProcess.getId(),
+			titleMap.get(contextAcceptLanguage.getPreferredLocale()), titleMap,
+			process.getVersion());
+	}
+
+	private BooleanQuery _createBooleanQuery(Long processId) {
 		BooleanQuery booleanQuery = _queries.booleanQuery();
 
-		return booleanQuery.addShouldQueryClauses(
-			_queries.term("completed", completed),
-			_queries.term("instanceId", 0));
-	}
-
-	private BooleanQuery _createBooleanQuery(
-		boolean completed, Date dateEnd, Date dateStart, Set<Long> processIds) {
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		booleanQuery.setMinimumShouldMatch(1);
-
-		BooleanQuery instancesBooleanQuery = _queries.booleanQuery();
-
-		instancesBooleanQuery.addFilterQueryClauses(
-			_queries.term(
-				"_index",
-				_instanceWorkflowMetricsIndexNameBuilder.getIndexName(
-					contextCompany.getCompanyId())));
-		instancesBooleanQuery.addMustNotQueryClauses(
-			_queries.term("instanceId", 0));
-		instancesBooleanQuery.addMustQueryClauses(
-			_createInstanceBooleanQuery(
-				completed, dateEnd, dateStart, processIds));
-
-		BooleanQuery slaInstanceResultsBooleanQuery = _queries.booleanQuery();
-
-		slaInstanceResultsBooleanQuery.addFilterQueryClauses(
-			_queries.term(
-				"_index",
-				_slaInstanceResultWorkflowMetricsIndexNameBuilder.getIndexName(
-					contextCompany.getCompanyId())));
-		slaInstanceResultsBooleanQuery.addMustNotQueryClauses(
-			_queries.term("slaDefinitionId", 0));
-		slaInstanceResultsBooleanQuery.addMustQueryClauses(
-			_createSLAInstanceResultsBooleanQuery(
-				completed, dateEnd, dateStart, processIds));
-
-		return booleanQuery.addShouldQueryClauses(
-			instancesBooleanQuery, slaInstanceResultsBooleanQuery);
-	}
-
-	private BucketSelectorPipelineAggregation
-		_createBucketSelectorPipelineAggregation() {
-
-		BucketSelectorPipelineAggregation bucketSelectorPipelineAggregation =
-			_aggregations.bucketSelector(
-				"bucketSelector", _scripts.script("params.instanceCount > 0"));
-
-		bucketSelectorPipelineAggregation.addBucketPath(
-			"instanceCount", "instanceCount.value");
-
-		return bucketSelectorPipelineAggregation;
-	}
-
-	private BooleanQuery _createCompletionDateBooleanQuery(
-		Date dateEnd, Date dateStart) {
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		return booleanQuery.addShouldQueryClauses(
-			_queries.rangeTerm(
-				"completionDate", true, true,
-				_resourceHelper.formatDate(dateStart),
-				_resourceHelper.formatDate(dateEnd)),
-			_queries.term("slaDefinitionId", 0));
-	}
-
-	private BooleanQuery _createInstanceBooleanQuery(
-		boolean completed, Date dateEnd, Date dateStart, Set<Long> processIds) {
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		if (completed && (dateEnd != null) && (dateStart != null)) {
-			booleanQuery.addMustQueryClauses(
-				_queries.rangeTerm(
-					"completionDate", true, true,
-					_resourceHelper.formatDate(dateStart),
-					_resourceHelper.formatDate(dateEnd)));
-		}
-
-		return booleanQuery.addMustQueryClauses(
-			_queries.term("companyId", contextCompany.getCompanyId()),
-			_queries.term("deleted", Boolean.FALSE),
-			_createBooleanQuery(completed),
-			_createProcessIdTermsQuery(processIds));
-	}
-
-	private Process _createProcess(Document document) {
-		return new Process() {
-			{
-				id = document.getLong("processId");
-				instanceCount = 0L;
-				onTimeInstanceCount = 0L;
-				overdueInstanceCount = 0L;
-				title = document.getString(_getTitleFieldName());
-			}
-		};
-	}
-
-	private BooleanQuery _createProcessBooleanQuery(
-		Long processId, String title) {
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		if (Validator.isNotNull(processId)) {
-			booleanQuery.addMustQueryClauses(
-				_queries.term("processId", processId));
-		}
-
-		if (Validator.isNotNull(title)) {
-			booleanQuery.addMustQueryClauses(_createTitleBooleanQuery(title));
-		}
+		booleanQuery.addMustQueryClauses(_queries.term("processId", processId));
 
 		return booleanQuery.addMustQueryClauses(
 			_queries.term("companyId", contextCompany.getCompanyId()),
 			_queries.term("deleted", Boolean.FALSE));
-	}
-
-	private TermsQuery _createProcessIdTermsQuery(Set<Long> processIds) {
-		TermsQuery termsQuery = _queries.terms("processId");
-
-		Stream<Long> stream = processIds.stream();
-
-		termsQuery.addValues(
-			stream.map(
-				String::valueOf
-			).toArray(
-				Object[]::new
-			));
-
-		return termsQuery;
-	}
-
-	private BooleanQuery _createSLAInstanceResultsBooleanQuery(
-		boolean completed, Date dateEnd, Date dateStart, Set<Long> processIds) {
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		if (completed) {
-			BooleanQuery shouldBooleanQuery = _queries.booleanQuery();
-
-			shouldBooleanQuery.addShouldQueryClauses(
-				_queries.term("slaDefinitionId", 0),
-				_queries.term("instanceCompleted", Boolean.TRUE));
-
-			booleanQuery.addMustQueryClauses(shouldBooleanQuery);
-
-			if ((dateEnd != null) && (dateStart != null)) {
-				booleanQuery.addMustQueryClauses(
-					_createCompletionDateBooleanQuery(dateEnd, dateStart));
-			}
-		}
-		else {
-			booleanQuery.addMustQueryClauses(
-				_queries.term("instanceCompleted", Boolean.FALSE));
-		}
-
-		return booleanQuery.addMustQueryClauses(
-			_queries.term("companyId", contextCompany.getCompanyId()),
-			_queries.term("deleted", Boolean.FALSE),
-			_createProcessIdTermsQuery(processIds));
-	}
-
-	private BooleanQuery _createTitleBooleanQuery(String title) {
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		StringQuery stringQuery = _queries.string(title + StringPool.STAR);
-
-		stringQuery.setDefaultField(_getTitleFieldName());
-
-		return booleanQuery.addShouldQueryClauses(
-			stringQuery, _queries.match(_getTitleFieldName(), title));
-	}
-
-	private TermsAggregationResult _getInstanceTermsAggregationResult(
-		boolean completed, FieldSort fieldSort, Pagination pagination,
-		Set<Long> processIds) {
-
-		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
-
-		TermsAggregation termsAggregation = _aggregations.terms(
-			"processId", "processId");
-
-		BooleanQuery booleanQuery = _queries.booleanQuery();
-
-		FilterAggregation filterAggregation = _aggregations.filter(
-			"instanceCountFilter",
-			booleanQuery.addMustNotQueryClauses(
-				_queries.term("instanceId", "0")));
-
-		filterAggregation.addChildAggregation(
-			_aggregations.valueCount("instanceCount", "instanceId"));
-
-		termsAggregation.addChildrenAggregations(filterAggregation);
-
-		if ((fieldSort != null) &&
-			_isOrderByInstanceCount(fieldSort.getField())) {
-
-			termsAggregation.addPipelineAggregation(
-				_resourceHelper.createBucketSortPipelineAggregation(
-					fieldSort, pagination));
-		}
-
-		termsAggregation.setSize(processIds.size());
-
-		searchSearchRequest.addAggregation(termsAggregation);
-
-		searchSearchRequest.setIndexNames(
-			_instanceWorkflowMetricsIndexNameBuilder.getIndexName(
-				contextCompany.getCompanyId()));
-		searchSearchRequest.setQuery(
-			_createInstanceBooleanQuery(completed, null, null, processIds));
-
-		SearchSearchResponse searchSearchResponse =
-			_searchRequestExecutor.executeSearchRequest(searchSearchRequest);
-
-		Map<String, AggregationResult> aggregationResultsMap =
-			searchSearchResponse.getAggregationResultsMap();
-
-		return (TermsAggregationResult)aggregationResultsMap.get("processId");
-	}
-
-	private Bucket _getProcessBucket(
-		boolean completed, Date dateEnd, Date dateStart, long processId) {
-
-		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
-
-		TermsAggregation termsAggregation = _aggregations.terms(
-			"processId", "processId");
-
-		FilterAggregation onTimeFilterAggregation = _aggregations.filter(
-			"onTime", _resourceHelper.createMustNotBooleanQuery());
-
-		onTimeFilterAggregation.addChildAggregation(
-			_resourceHelper.createOnTimeScriptedMetricAggregation());
-
-		FilterAggregation overdueFilterAggregation = _aggregations.filter(
-			"overdue", _resourceHelper.createMustNotBooleanQuery());
-
-		overdueFilterAggregation.addChildAggregation(
-			_resourceHelper.createOverdueScriptedMetricAggregation());
-
-		termsAggregation.addChildrenAggregations(
-			onTimeFilterAggregation, overdueFilterAggregation,
-			_resourceHelper.creatInstanceCountScriptedMetricAggregation(
-				Collections.emptyList(), dateEnd, dateStart,
-				Collections.emptyList(), Collections.emptyList(),
-				Collections.emptyList()));
-
-		termsAggregation.addPipelineAggregations(
-			_createBucketSelectorPipelineAggregation());
-
-		searchSearchRequest.addAggregation(termsAggregation);
-
-		searchSearchRequest.setIndexNames(
-			_instanceWorkflowMetricsIndexNameBuilder.getIndexName(
-				contextCompany.getCompanyId()),
-			_slaInstanceResultWorkflowMetricsIndexNameBuilder.getIndexName(
-				contextCompany.getCompanyId()));
-
-		searchSearchRequest.setQuery(
-			_createBooleanQuery(
-				completed, dateEnd, dateStart,
-				Collections.singleton(processId)));
-
-		SearchSearchResponse searchSearchResponse =
-			_searchRequestExecutor.executeSearchRequest(searchSearchRequest);
-
-		Map<String, AggregationResult> aggregationResultsMap =
-			searchSearchResponse.getAggregationResultsMap();
-
-		TermsAggregationResult termsAggregationResult =
-			(TermsAggregationResult)aggregationResultsMap.get("processId");
-
-		return termsAggregationResult.getBucket(String.valueOf(processId));
-	}
-
-	private Collection<Process> _getProcesses(
-			FieldSort fieldSort, Pagination pagination, SearchHits searchHits)
-		throws Exception {
-
-		List<Process> processes = new LinkedList<>();
-
-		Map<Long, Process> processesMap = Stream.of(
-			searchHits.getSearchHits()
-		).flatMap(
-			List::stream
-		).map(
-			SearchHit::getDocument
-		).map(
-			this::_createProcess
-		).collect(
-			LinkedHashMap::new,
-			(map, process) -> map.put(process.getId(), process), Map::putAll
-		);
-
-		TermsAggregationResult instanceTermsAggregationResult =
-			_getInstanceTermsAggregationResult(
-				false, fieldSort, pagination, processesMap.keySet());
-		TermsAggregationResult slaTermsAggregationResult =
-			_getSLATermsAggregationResult(
-				false, fieldSort, pagination, processesMap.keySet());
-
-		if (_isOrderByInstanceCount(fieldSort.getField())) {
-			for (Bucket bucket : instanceTermsAggregationResult.getBuckets()) {
-				Process process = processesMap.remove(
-					Long.valueOf(bucket.getKey()));
-
-				_populateProcessWithSLAMetrics(
-					slaTermsAggregationResult.getBucket(bucket.getKey()),
-					process);
-				_setInstanceCount(bucket, process);
-				_setUntrackedInstanceCount(process);
-
-				processes.add(process);
-			}
-		}
-		else if (_isOrderByTitle(fieldSort.getField())) {
-			for (Process process : processesMap.values()) {
-				_populateProcessWithSLAMetrics(
-					slaTermsAggregationResult.getBucket(
-						String.valueOf(process.getId())),
-					process);
-				_setInstanceCount(
-					instanceTermsAggregationResult.getBucket(
-						String.valueOf(process.getId())),
-					process);
-				_setUntrackedInstanceCount(process);
-
-				processes.add(process);
-			}
-		}
-		else {
-			for (Bucket bucket : slaTermsAggregationResult.getBuckets()) {
-				Process process = processesMap.remove(
-					Long.valueOf(bucket.getKey()));
-
-				_populateProcessWithSLAMetrics(bucket, process);
-				_setInstanceCount(
-					instanceTermsAggregationResult.getBucket(bucket.getKey()),
-					process);
-				_setUntrackedInstanceCount(process);
-
-				processes.add(process);
-			}
-		}
-
-		return processes;
-	}
-
-	private SearchSearchResponse _getProcessesSearchSearchResponse(
-		FieldSort fieldSort, Pagination pagination, Long processId,
-		String title) {
-
-		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
-
-		searchSearchRequest.setIndexNames(
-			_processWorkflowMetricsIndexNameBuilder.getIndexName(
-				contextCompany.getCompanyId()));
-		searchSearchRequest.setQuery(
-			_createProcessBooleanQuery(processId, title));
-		searchSearchRequest.setSelectedFieldNames(
-			"processId", _getTitleFieldName());
-
-		if ((fieldSort != null) && _isOrderByTitle(fieldSort.getField())) {
-			searchSearchRequest.setSize(pagination.getPageSize());
-			searchSearchRequest.setSorts(Collections.singletonList(fieldSort));
-			searchSearchRequest.setStart(pagination.getStartPosition());
-		}
-		else {
-			searchSearchRequest.setSize(10000);
-			searchSearchRequest.setStart(0);
-		}
-
-		return _searchRequestExecutor.executeSearchRequest(searchSearchRequest);
-	}
-
-	private TermsAggregationResult _getSLATermsAggregationResult(
-		boolean completed, FieldSort fieldSort, Pagination pagination,
-		Set<Long> processIds) {
-
-		SearchSearchRequest searchSearchRequest = new SearchSearchRequest();
-
-		TermsAggregation termsAggregation = _aggregations.terms(
-			"processId", "processId");
-
-		FilterAggregation onTimeFilterAggregation = _aggregations.filter(
-			"onTime", _resourceHelper.createMustNotBooleanQuery());
-
-		onTimeFilterAggregation.addChildAggregation(
-			_resourceHelper.createOnTimeScriptedMetricAggregation());
-
-		FilterAggregation overdueFilterAggregation = _aggregations.filter(
-			"overdue", _resourceHelper.createMustNotBooleanQuery());
-
-		overdueFilterAggregation.addChildAggregation(
-			_resourceHelper.createOverdueScriptedMetricAggregation());
-
-		termsAggregation.addChildrenAggregations(
-			onTimeFilterAggregation, overdueFilterAggregation);
-
-		if ((fieldSort != null) &&
-			!_isOrderByInstanceCount(fieldSort.getField()) &&
-			!_isOrderByTitle(fieldSort.getField())) {
-
-			termsAggregation.addPipelineAggregation(
-				_resourceHelper.createBucketSortPipelineAggregation(
-					fieldSort, pagination));
-		}
-
-		termsAggregation.setSize(processIds.size());
-
-		searchSearchRequest.addAggregation(termsAggregation);
-
-		searchSearchRequest.setIndexNames(
-			_slaInstanceResultWorkflowMetricsIndexNameBuilder.getIndexName(
-				contextCompany.getCompanyId()));
-		searchSearchRequest.setQuery(
-			_createSLAInstanceResultsBooleanQuery(
-				completed, null, null, processIds));
-
-		SearchSearchResponse searchSearchResponse =
-			_searchRequestExecutor.executeSearchRequest(searchSearchRequest);
-
-		Map<String, AggregationResult> aggregationResultsMap =
-			searchSearchResponse.getAggregationResultsMap();
-
-		return (TermsAggregationResult)aggregationResultsMap.get("processId");
 	}
 
 	private String _getTitleFieldName() {
@@ -607,121 +158,8 @@ public class ProcessResourceImpl
 			contextAcceptLanguage.getPreferredLocale(), "title");
 	}
 
-	private boolean _isOrderByInstanceCount(String fieldName) {
-		return StringUtil.startsWith(fieldName, "instanceCount");
-	}
-
-	private boolean _isOrderByTitle(String fieldName) {
-		return StringUtil.startsWith(fieldName, "title");
-	}
-
-	private void _populateProcessWithSLAMetrics(
-		Bucket bucket, Process process) {
-
-		_setOnTimeInstanceCount(bucket, process);
-		_setOverdueInstanceCount(bucket, process);
-	}
-
-	private void _setInstanceCount(Bucket bucket, Process process) {
-		if (bucket == null) {
-			return;
-		}
-
-		FilterAggregationResult filterAggregationResult =
-			(FilterAggregationResult)bucket.getChildAggregationResult(
-				"instanceCountFilter");
-
-		ValueCountAggregationResult valueCountAggregationResult;
-
-		if (filterAggregationResult != null) {
-			valueCountAggregationResult =
-				(ValueCountAggregationResult)
-					filterAggregationResult.getChildAggregationResult(
-						"instanceCount");
-
-			process.setInstanceCount(valueCountAggregationResult.getValue());
-		}
-		else {
-			ScriptedMetricAggregationResult scriptedMetricAggregationResult =
-				(ScriptedMetricAggregationResult)
-					bucket.getChildAggregationResult("instanceCount");
-
-			process.setInstanceCount(
-				GetterUtil.getLong(scriptedMetricAggregationResult.getValue()));
-		}
-	}
-
-	private void _setOnTimeInstanceCount(Bucket bucket, Process process) {
-		if (bucket == null) {
-			return;
-		}
-
-		process.setOnTimeInstanceCount(
-			_resourceHelper.getOnTimeInstanceCount(bucket));
-	}
-
-	private void _setOverdueInstanceCount(Bucket bucket, Process process) {
-		if (bucket == null) {
-			return;
-		}
-
-		process.setOverdueInstanceCount(
-			_resourceHelper.getOverdueInstanceCount(bucket));
-	}
-
-	private void _setUntrackedInstanceCount(Process process) {
-		long untrackedInstanceCount =
-			process.getInstanceCount() -
-				GetterUtil.getLong(process.getOnTimeInstanceCount()) -
-					GetterUtil.getLong(process.getOverdueInstanceCount());
-
-		if (untrackedInstanceCount < 0) {
-			untrackedInstanceCount = 0;
-		}
-
-		process.setUntrackedInstanceCount(untrackedInstanceCount);
-	}
-
-	private FieldSort _toFieldSort(Sort[] sorts) {
-		String titleFieldName = Field.getSortableFieldName(
-			_getTitleFieldName());
-
-		Sort sort = new Sort(titleFieldName, false);
-
-		if (sorts != null) {
-			sort = sorts[0];
-		}
-
-		String fieldName = sort.getFieldName();
-
-		if (_isOrderByInstanceCount(fieldName)) {
-			fieldName = "instanceCountFilter > instanceCount";
-		}
-		else if (_isOrderByTitle(fieldName)) {
-			fieldName = titleFieldName;
-		}
-		else {
-			fieldName =
-				StringUtil.extractFirst(fieldName, "InstanceCount") +
-					" > instanceCount.value";
-		}
-
-		FieldSort fieldSort = _sorts.field(fieldName);
-
-		fieldSort.setSortOrder(
-			sort.isReverse() ? SortOrder.DESC : SortOrder.ASC);
-
-		return fieldSort;
-	}
-
-	private static final EntityModel _entityModel = new ProcessEntityModel();
-
 	@Reference
-	private Aggregations _aggregations;
-
-	@Reference(target = "(workflow.metrics.index.entity.name=instance)")
-	private WorkflowMetricsIndexNameBuilder
-		_instanceWorkflowMetricsIndexNameBuilder;
+	private ProcessWorkflowMetricsIndexer _processWorkflowMetricsIndexer;
 
 	@Reference(target = "(workflow.metrics.index.entity.name=process)")
 	private WorkflowMetricsIndexNameBuilder
@@ -731,21 +169,6 @@ public class ProcessResourceImpl
 	private Queries _queries;
 
 	@Reference
-	private ResourceHelper _resourceHelper;
-
-	@Reference
-	private Scripts _scripts;
-
-	@Reference
 	private SearchRequestExecutor _searchRequestExecutor;
-
-	@Reference(
-		target = "(workflow.metrics.index.entity.name=sla-instance-result)"
-	)
-	private WorkflowMetricsIndexNameBuilder
-		_slaInstanceResultWorkflowMetricsIndexNameBuilder;
-
-	@Reference
-	private Sorts _sorts;
 
 }
